requirements.txt
asgiref==3.5.2
attrs==22.2.0
black==22.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.4
django-allauth==0.51.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==16.1.0
idna==3.4
iniconfig==1.1.1
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==22.0
pathspec==0.10.3
platformdirs==2.6.0
pluggy==1.0.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.6.0
pytest==7.2.0
pytest-django==4.5.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.13
